io_u: use hw assisted random filling for buffers
If we're using refill_buffers and not doing any specific dedupe or
verification that requires a persistent buffer via a seed, then we
can just use hw assisted crc32c for a nice speedup here.

This bumps perf by about 2.4x for me in fill rate on aarch64, and
about a 28% improvement on x86-64 (using a 12900K).

Signed-off-by: Jens Axboe <axboe@kernel.dk>